feat: Require status checks
We only enable this for a specific set of repos: those which rely on crates
from eclipse-zenoh/zenoh. This is because the other repos do not yet have a
suitable CI job to be used as a status check (in GitHub status checks
correspond to job names, and as CI workflow jobs can change on a regular basis,
it would be quite brittle to list them here).

This will be used to automatically sync Cargo lockfiles between repos which
contain Rust crates that depend on eclipse-zenoh/zenoh. We plan on extending
the branch protections to all repos once their CI workflows are refactored.