Update Otterdog configuration (#5)
* chore: Remove unused secrets

The `PAT` secret was used to automate tasks not possible with `GITHUB_TOKEN`. Now that we have `BOT_TOKEN_WORKFLOW` this is no longer necessary, and it simply polluting the config.

* chore: Enable auto-merge everywhere

GitHub's auto-merge feature is useful as it allows for automatically creating pull requests that wait for status checks to pass before being automatically merged.

* chore: Change default branch to 'main' on all repos

* fix: Set default branch name on org level

* feat: Require status checks

We only enable this for a specific set of repos: those which rely on crates
from eclipse-zenoh/zenoh. This is because the other repos do not yet have a
suitable CI job to be used as a status check (in GitHub status checks
correspond to job names, and as CI workflow jobs can change on a regular basis,
it would be quite brittle to list them here).

This will be used to automatically sync Cargo lockfiles between repos which
contain Rust crates that depend on eclipse-zenoh/zenoh. We plan on extending
the branch protections to all repos once their CI workflows are refactored.

* fix: Correct branch protection rules syntax

* Convert branch protection rules to rulesets

Signed-off-by: Thomas Neidhart <thomas.neidhart@eclipse-foundation.org>

---------

Signed-off-by: Thomas Neidhart <thomas.neidhart@eclipse-foundation.org>
Co-authored-by: Thomas Neidhart <thomas.neidhart@eclipse-foundation.org>